+++ flutter_lwa/android/src/main/java/com/github/ayvazj/flutter_lwa/ProofKeyParameters.java
@@ -1,0 +1,12 @@
+package com.github.ayvazj.flutter_lwa;
+import androidx.annotation.NonNull;
+
+final class ProofKeyParameters {
+    public final String codeChallenge;
+    public final String codeChallengeMethod;
+
+    ProofKeyParameters(@NonNull String codeChallenge, @NonNull String codeChallengeMethod) {
+        this.codeChallenge = codeChallenge;
+        this.codeChallengeMethod = codeChallengeMethod;
+    }
+}