+++ flutter_lwa/ios/Classes/LwaPlugin.m
@@ -2,6 +2,7 @@
 
 
 #import <LoginWithAmazon/LoginWithAmazon.h>
+#import "ProofKeyParameters.h"
 
 @implementation LwaPlugin
 NSString *const kMethodSignIn = @"SIGNIN";
@@ -9,7 +10,12 @@ NSString *const kMethodSignOut = @"SIGNOUT";
 NSString *const kMethodGetToken = @"GETTOKEN";
 NSString *const kMethodGetProfile = @"GETPROFILE";
 NSString *const kScopesArgument = @"scopes";
-
+NSString *const kGrantTypeArgument = @"grantType";
+NSString *const kProofKeyParametersArgument = @"proofKeyParameters";
+NSString *const kCodeChallengeArgument = @"codeChallenge";
+NSString *const kCodeChallengeMethodArgument = @"codeChallengeMethod";
+NSString *const kGrantTypeAccessTokenArgument = @"ACCESS_TOKEN";
+NSString *const kGrantTypeAccessAuthorizationCodeArgument = @"AUTHORIZATION_CODE";
 NSString *const kProfileScopeProfile = @"profile";
 NSString *const kProfileScopePostalCode = @"postal_code";
 NSString *const kProfileScopeUserId = @"user_id";
@@ -40,7 +46,7 @@ NSDictionary *kDefinedScopes;
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
     if ([kMethodSignIn isEqualToString:call.method.uppercaseString]) {
-        [self.delegate signIn:result scopes:[LwaPlugin getScopesArgument:call]];
+        [self.delegate signIn:result scopes:[LwaPlugin getScopesArgument:call] grantType:[LwaPlugin getGrantTypeArgument:call] proofKeyParameters:[LwaPlugin getProofKeyParametersArgument:call]];
     } else if ([kMethodSignOut isEqualToString:call.method.uppercaseString]) {
         [self.delegate signOut:result];
     } else if ([kMethodGetToken isEqualToString:call.method.uppercaseString]) {
@@ -70,6 +76,43 @@ NSDictionary *kDefinedScopes;
         }
     }
     return [NSArray arrayWithArray:res];
+}
+
++ (AMZNAuthorizationGrantType) getGrantTypeArgument:(FlutterMethodCall*)call {
+    NSString* grantTypeArgument = call.arguments[kGrantTypeArgument];
+    if (grantTypeArgument) {
+        if ([grantTypeArgument caseInsensitiveCompare:kGrantTypeAccessTokenArgument] == NSOrderedSame) {
+            return AMZNAuthorizationGrantTypeToken;
+        }
+        if ([grantTypeArgument caseInsensitiveCompare:kGrantTypeAccessAuthorizationCodeArgument] == NSOrderedSame) {
+            return AMZNAuthorizationGrantTypeCode;
+        }
+    }
+    return AMZNAuthorizationGrantTypeToken; // default is AMZNAuthorizationGrantTypeToken
+}
+
++ (ProofKeyParameters*) getProofKeyParametersArgument:(FlutterMethodCall*)call {
+    NSDictionary* proofKeyParametersArgument = call.arguments[kProofKeyParametersArgument];
+    NSString *codeChallenge = nil;
+    NSString *codeChallengeMethod = nil;
+    if (proofKeyParametersArgument != nil) {
+        NSArray *allKeys = [proofKeyParametersArgument allKeys];
+        for (NSString *str in allKeys) {
+            if ([kCodeChallengeArgument caseInsensitiveCompare:str] == NSOrderedSame) {
+                codeChallenge = [proofKeyParametersArgument objectForKey:str];
+            }
+            if ([kCodeChallengeMethodArgument caseInsensitiveCompare:str] == NSOrderedSame) {
+                codeChallengeMethod = [proofKeyParametersArgument objectForKey:str];
+            }
+        }
+        if (codeChallenge != nil && codeChallengeMethod != nil) {
+            ProofKeyParameters* res = [[ProofKeyParameters alloc] init];
+            res.codeChallenge = codeChallenge;
+            res.codeChallengeMethod = codeChallengeMethod;
+            return res;
+        }
+    }
+    return nil;
 }
 
 @end