+++ flutter_lwa/ios/Classes/ProofKeyParameters.h
@@ -1,0 +1,10 @@
+#ifndef ProofKeyParameters_h
+#define ProofKeyParameters_h
+
+@interface ProofKeyParameters : NSObject
+@property NSString *codeChallenge;
+@property NSString *codeChallengeMethod;
+@end
+
+#endif /* ProofKeyParameters_h */
+