+++ flutter_lwa/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0
+
+* Implemented GrantType and ProofKeyParameters arguments to enable using Alexa APIs
+* Updated example with options for standard LWA and AVS authentication.
+
 ## 2.0.0
 
 * Refactor to support federated development.